dev/backend/requirements.txt
Flask
Flask-CORS
Flask-Mail
mysql-connector
bcrypt